+++ Source/FindProgramsViewController.swift
@@ -17,6 +17,7 @@ class FindProgramsViewController: UIViewController, InterfaceOrientationOverridi
     fileprivate var showBottomBar: Bool = true
     fileprivate let searchQuery: String?
     fileprivate let bottomBar: UIView?
+    private(set) var pathId: String?
     private(set) var webviewHelper: DiscoveryWebViewHelper?
     private var enrollmentConfig: EnrollmentConfig? {
         return environment.config.programEnrollment
@@ -35,6 +36,11 @@ class FindProgramsViewController: UIViewController, InterfaceOrientationOverridi
         self.showBottomBar = showBottomBar
     }
     
+    convenience init(with environment: Environment, pathId: String, bottomBar: UIView?) {
+        self.init(with: environment, bottomBar: bottomBar, searchQuery: nil)
+        self.pathId = pathId
+    }
+    
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
@@ -43,10 +49,21 @@ class FindProgramsViewController: UIViewController, InterfaceOrientationOverridi
     override func viewDidLoad() {
         super.viewDidLoad()
         navigationItem.title = Strings.discover
-        webviewHelper = DiscoveryWebViewHelper(environment: environment, delegate: self, bottomBar: showBottomBar ? bottomBar : nil, showSearch: true, searchQuery: searchQuery, discoveryType: .programs)
-        if let url = enrollmentConfig?.webview.searchURL as URL? {
-            webviewHelper?.searchBaseURL = url
-            webviewHelper?.load(withURL: url)
+        if let pathId = pathId {
+            addBackBarButton()
+            webviewHelper = DiscoveryWebViewHelper(environment: environment, delegate: self, bottomBar: bottomBar, showSearch: false, searchQuery: nil, discoveryType: .programs)
+            if let detailTemplate = enrollmentConfig?.webview.detailTemplate?.replacingOccurrences(of: URIString.pathPlaceHolder.rawValue, with: pathId),
+                let url = URL(string: detailTemplate) {
+                webviewHelper?.load(withURL: url)
+            }
+            
+        }
+        else {
+            webviewHelper = DiscoveryWebViewHelper(environment: environment, delegate: self, bottomBar: showBottomBar ? bottomBar : nil, showSearch: true, searchQuery: searchQuery, discoveryType: .programs)
+            if let url = enrollmentConfig?.webview.searchURL as URL? {
+                webviewHelper?.searchBaseURL = url
+                webviewHelper?.load(withURL: url)
+            }
         }
         
     }
